refactor: authController from authRoutes
